TortoiseBlame's context menu now has "Blame previous revision" which makes tracking changes easier. See issue #343.
Calling "Blame Previous Revision" on a selected revision you will get the blame on the previous revision. The "From revision" is determined automatically based on the revisions which are shown in the current blame: the smallest revision number is used as the "From revision". TortoiseBlame.exe has a new optional parameter /path:originalpath. This is needed so that TortoiseBlame knows which path to pass on to TortoiseProc.exe.